% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_df.r
\name{describe_df}
\alias{describe_df}
\title{Quick description of dataframe}
\usage{
describe_df(data)
}
\arguments{
\item{data}{dataframe to describe}
}
\value{
A new dataframe with informations about the input dataframe
}
\description{
This functions returns a dataframe with some information:
name of variables, labels, number of factor levels...
}
