% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ow.r
\name{ow}
\alias{ow}
\title{Oneway table}
\usage{
ow(
  data,
  var,
  rounding_prct = 2,
  hide_prct_char = TRUE,
  no_na_prct = TRUE,
  cond_prct = FALSE,
  min_cond_prct = 0.05,
  cond_excluded_labels = NULL,
  lang = "en"
)
}
\arguments{
\item{data}{Dataframe to use}

\item{var}{Variable to use}

\item{rounding_prct}{number of digits for rounding percentages (int)}

\item{hide_prct_char}{hides the "\%" characted in the percentages column}

\item{no_na_prct}{Removes NAs from \%ages calculations}

\item{cond_prct}{create a new percentage column on some specific, with \% computed conditionally on a threshold (bool)}

\item{min_cond_prct}{Threshold to compute the new \%s (int)}

\item{cond_excluded_labels}{Restrict the exclusion of \%s calculations to specific value labels (list)}

\item{lang}{changes the name of the percentage column (depending on your language). Three values : "en" (default), "fr", or "math"}
}
\value{
Count table
}
\description{
This functions returns a oneway table of a variable
}
