% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ow_report.r
\name{ow_report}
\alias{ow_report}
\title{Oneway report of multiple categorical variables to Excel}
\usage{
ow_report(
  data,
  workbook,
  worksheet,
  vars,
  rounding = 2,
  total = TRUE,
  filename
)
}
\arguments{
\item{workbook}{name of the workbook (string)}

\item{worksheet}{name of the worksheet (string)}

\item{vars}{Variables to report (string)}

\item{rounding}{number of digits for rounding (int)}

\item{filename}{Name of excel file to export to}

\item{Data}{dataframe}

\item{total_col}{add a total column (bool)}
}
\value{
Excel file with oneway table
}
\description{
This functions creates an Excel workbook and exports a oneway table
of multiple variables with counts and percentages
}
