% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_message.r
\name{pipe_message}
\alias{pipe_message}
\title{Display a message mid-chain}
\usage{
pipe_message(.data, status)
}
\arguments{
\item{.data}{Input data from chaining}

\item{status}{Message to display}
}
\value{
Input dataframe and a message
}
\description{
Stolen from: https://stackoverflow.com/questions/46123285/printing-intermediate-results-without-breaking-pipeline-in-tidyverse
}
