% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_print.r
\name{pipe_print}
\alias{pipe_print}
\title{Print an object mid-chain}
\usage{
pipe_print(.data, content)
}
\arguments{
\item{.data}{Input data from chaining}

\item{content}{Content to display using print (e.g. a message, a table...)}
}
\value{
Input dataframe and a message
}
\description{
Stolen from: https://stackoverflow.com/questions/46123285/printing-intermediate-results-without-breaking-pipeline-in-tidyverse
}
