% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_label.r
\name{put_label}
\alias{put_label}
\title{Replace a variable name (var) with a label in another dataframe}
\usage{
put_label(data, var, data_label, label_from, label_to)
}
\arguments{
\item{data}{dataframe}

\item{var}{Variable to rename}

\item{data_label}{dataframe with the list of variables in data (in one column), and their label (in another column)}

\item{label_from}{column in data_label of variables from data}

\item{label_to}{column in data_label of labels}
}
\value{
Same table, with a renamed variable
}
\description{
In oneway or twoway tables, replace the name of the variable with its label
The label must be stored in another dataframe, with: a column with the list of variables
and another column with its associated label
}
