% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_group_report.r
\name{quant_group_report}
\alias{quant_group_report}
\title{Report of multiple quantitative variables by group to Excel}
\usage{
quant_group_report(
  data,
  workbook,
  worksheet,
  groupvar,
  quantvars,
  rounding = 2,
  filename
)
}
\arguments{
\item{workbook}{name of the workbook (string)}

\item{worksheet}{name of the worksheet (string)}

\item{quantvars}{names of row variables (vector of strings)}

\item{rounding}{number of digits for rounding}

\item{filename}{Name of excel file to export to}

\item{Data}{dataframe}

\item{colvar}{name of column variable (string, max one)}

\item{percentage_style}{"row", "col" or "both" (= cell) percentages}

\item{counts}{show counts in parenthesis}

\item{separate_counts}{show counts in a new column}

\item{total_col}{add a total column (bool)}

\item{pval_col}{put the p-values in a new column instead of new rows (bool)}

\item{chi_cat_reject}{Automatic drop of categories with low counts}

\item{chi_cat_reject_threshold}{Threshold at which categories are dropped (default:  categories with n <= 30 are dropped). Use a value >= to 1 to drop based on counts (n), or < 1 to drop based on percentages (e.g. 0.01)}

\item{show_chi_table}{Returns the table used to compute the chi-squared test}
}
\value{
Excel file with twoway table
}
\description{
This functions creates an Excel workbook and exports a twoway table
of one column variable and multiple row variables, with chi-squared tests.
}
