% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_report.r
\name{quant_report}
\alias{quant_report}
\title{Descriptive statistics for continuous variables}
\usage{
quant_report(
  data,
  workbook,
  worksheet,
  vars,
  filename,
  open_table_on_completion = FALSE
)
}
\arguments{
\item{data}{Dataframe to use}

\item{columns}{String of list of strings with columns to describe}
}
\value{
A dataframe with descriptive statistics
}
\description{
This functions returns a dataframe with some descriptive statistics from continuous vars
name of variables, labels, number of factor levels...
}
