% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_summary.r
\name{quant_summary}
\alias{quant_summary}
\title{Descriptive statistics for continuous variables}
\usage{
quant_summary(data, columns)
}
\arguments{
\item{data}{Dataframe to use}

\item{columns}{String of list of strings with columns to describe}
}
\value{
A dataframe with descriptive statistics
}
\description{
This functions returns a dataframe with some descriptive statistics from continuous vars
name of variables, labels, number of factor levels...
}
