% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_report.r
\name{regression_report}
\alias{regression_report}
\title{Regression table report to XL}
\usage{
regression_report(
  data,
  workbook,
  new_wb = TRUE,
  worksheet,
  model_label = "Model",
  lang = "en",
  depvar,
  data_label,
  label_from,
  label_to,
  open_on_finish = TRUE,
  overwrite_file = TRUE,
  filename
)
}
\arguments{
\item{data}{Dataframe to use}

\item{workbook}{name of the workbook (string)}

\item{new_wb}{create new workbook (or open filename)}

\item{worksheet}{name of the worksheet (string)}

\item{lang}{changes the name of the percentage column (depending on your language). Three values : "en" (default), "fr", or "math"}

\item{depvar}{Dependent variable name (used to retrieve its label)}

\item{data_label}{Dataframe of value labels from which to retrieve variable labels (optionnal)}

\item{label_from}{Column (in data_label) of variable names}

\item{label_to}{Column (in data_label) of variable labels}

\item{open_on_finish}{open excel file on finish (bool)}

\item{overwrite_file}{Overwrite existing file (bool)}

\item{filename}{Name of excel file to export to (string)}
}
\value{
Excel file with twoway table
}
\description{
This functions creates an Excel workbook and exports a regression table
}
