% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_tidy_results.r
\name{regression_tidy_results}
\alias{regression_tidy_results}
\title{Tidy results to use with regression_report}
\usage{
regression_tidy_results(
  data,
  exponentiate = FALSE,
  robust = FALSE,
  robust_method = "HC0"
)
}
\arguments{
\item{data}{Dataframe to use}

\item{robust}{Whether to include robust results (uses lmtest package)}

\item{robust_method}{Which robust method to use (see vcovHC in sandwich package)}
}
\value{
Formatted regression table to use with regression_report
}
\description{
This function tidies the raw output of some regressions using
broom, shows robust estimates if required, and does some
formatting so the output can be used with regression_report()
}
