% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{report}
\alias{report}
\title{Export report (for tw tables only for now)}
\usage{
report(
  data,
  workbook,
  worksheet,
  label_data,
  label_from,
  label_to,
  open_on_finish = FALSE,
  append_to_existing_file = FALSE,
  filename
)
}
\arguments{
\item{data}{Dataframe to use}

\item{workbook}{name of the workbook (string)}

\item{worksheet}{name of the worksheet (string)}

\item{label_data}{Dataframe of value labels from which to retrieve variable labels (optional)}

\item{label_from}{Column (in data_label) of variable names}

\item{label_to}{Column (in data_label) of variable labels}

\item{open_on_finish}{open excel file on finish (bool)}

\item{append_to_existing_file}{Append report ton an existing excel file}

\item{filename}{Name of excel file to export to (string)}
}
\value{
An excel file + the original table
}
\description{
This functions exports an excel file of a tw table
}
