% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_bygroup_mean.r
\name{svy_bygroup_mean}
\alias{svy_bygroup_mean}
\title{Oneway report of multiple categorical variables to Excel}
\usage{
svy_bygroup_mean(
  data,
  quant_var,
  by_var,
  rounding_prct = 2,
  rounding_ci = 2,
  show_se = FALSE,
  show_ci = TRUE
)
}
\arguments{
\item{data}{Dataframe to use}

\item{quant_var}{Quantitative variable to use}

\item{by_var}{Variable to use for group_by}

\item{rounding_prct}{number of digits for rounding percentages (int)}

\item{rounding_ci}{hides the "\%" characted in the percentages column}

\item{show_se}{shows standard error of the mean}

\item{show_ci}{shows the confidence interval of the mean}
}
\value{
Table with means, SE, CI and p-values
}
\description{
This functions returns a oneway table of a variable
Must be used on a tbl_svy dataframe
}
