% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_bygroup_mean_report.r
\name{svy_bygroup_mean_report}
\alias{svy_bygroup_mean_report}
\title{Report of a quantitative variable by multiple groups, with t-tests}
\usage{
svy_bygroup_mean_report(
  data,
  workbook,
  new_wb = TRUE,
  worksheet,
  quant_var,
  by_vars,
  rounding_mean = 2,
  rounding_se = 2,
  rounding_prct = 2,
  rounding_ci = 2,
  show_se = FALSE,
  show_ci = TRUE,
  lang = "en",
  data_label,
  label_from,
  label_to,
  open_on_finish = FALSE,
  overwrite_file = TRUE,
  filename
)
}
\arguments{
\item{data}{Dataframe to use}

\item{workbook}{name of the workbook (string)}

\item{new_wb}{create new workbook (or open filename)}

\item{worksheet}{name of the worksheet (string)}

\item{quant_var}{Quantitative variable to use}

\item{by_vars}{Variable to use for group_by}

\item{rounding_mean}{number of digits for rounding mean (int)}

\item{rounding_se}{number of digits for rounding mean standard error (int)}

\item{rounding_prct}{number of digits for rounding percentages (int)}

\item{rounding_ci}{hides the "\%" characted in the percentages column}

\item{show_se}{shows standard error of the mean}

\item{show_ci}{shows the confidence interval of the mean}

\item{lang}{changes some presentation texts (depending on your language). Three values : "en" (default), "fr", or "math"}

\item{data_label}{Dataframe of value labels from which to retrieve variable labels (optionnal)}

\item{label_from}{Column (in data_label) of variable names}

\item{label_to}{Column (in data_label) of variable labels}

\item{open_on_finish}{open excel file on finish (bool)}

\item{overwrite_file}{Overwrite existing file (bool)}

\item{filename}{Name of excel file to export to (string)}
}
\value{
Table with means, SE, CI and p-values
}
\description{
This functions returns a by-group summary of means for a continuous variable (with t-test of difference in means using a regression)
Must be used on a tbl_svy dataframe
}
