% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_ow.r
\name{svy_ow}
\alias{svy_ow}
\title{Oneway report of multiple categorical variables to Excel}
\usage{
svy_ow(data, var, rounding = 2, hide_prct_char = TRUE, lang = "en")
}
\arguments{
\item{data}{Dataframe to use}

\item{var}{Variable to use}

\item{rounding}{number of digits for rounding (int)}

\item{hide_prct_char}{hides the "\%" characted in the percentages column}

\item{lang}{changes the name of the percentage column (depending on your language). Three values : "en" (default), "fr", or "math"}
}
\value{
Count table
}
\description{
This functions returns a oneway table of a variable
Must be used on a tbl_svy dataframe
}
