% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_ow_report.r
\name{svy_ow_report}
\alias{svy_ow_report}
\title{Oneway report of multiple categorical variables to Excel, with survey design}
\usage{
svy_ow_report(
  data,
  workbook,
  worksheet,
  vars,
  rounding = 2,
  data_label,
  label_from,
  label_to,
  open_on_finish = TRUE,
  overwrite_file = TRUE,
  lang = "en",
  filename
)
}
\arguments{
\item{workbook}{name of the workbook (string)}

\item{worksheet}{name of the worksheet (string)}

\item{vars}{Variables to report (string)}

\item{rounding}{number of digits for rounding (int)}

\item{open_on_finish}{open excel file on finish (bool)}

\item{overwrite_file}{Overwrite existing file (bool)}

\item{lang}{Language of header (one of "en", "fr" or "math")}

\item{filename}{Name of excel file to export to (string)}

\item{Data}{dataframe}
}
\value{
Excel file with oneway table
}
\description{
This functions creates an Excel workbook and exports a oneway table
of multiple variables with counts and percentages (with a survey design)
}
