% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw.r
\name{tw}
\alias{tw}
\title{Twoway table}
\usage{
tw(
  data,
  colvar,
  rowvars,
  type_prct = "col",
  add_row_total = FALSE,
  add_col_total = FALSE,
  rounding_prct = 2,
  missing = TRUE,
  chisq_n_thres = 5,
  chisq_show_excluded = TRUE,
  lang = "en"
)
}
\arguments{
\item{data}{Dataframe to use}

\item{colvar}{Variable to use as the column variable}

\item{type_prct}{Type of percentage shown (one of "col", "row" or "cell")}

\item{add_row_total}{Add total column for row variable}

\item{rounding_prct}{number of digits for rounding percentages (int)}

\item{missing}{include NAs from the table (bool)}

\item{chisq_n_thres}{Minimum n for chi-squared test, if >= 1. Minimum percentage for chi-squared test, if < 1}

\item{chisq_show_excluded}{Show which variables have excluded categories}

\item{lang}{changes the name of the percentage column (depending on your language). Three values : "en" (default), "fr", or "math"}

\item{rowvar}{list of variables to use as row variables}
}
\value{
Tibble with one or multiple two-way tables
}
\description{
This functions returns a tibble with one or multiple two-way tables between one column variable and multiple row variables
}
