% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_ow.r
\name{svy_ow}
\alias{svy_ow}
\title{Oneway report of multiple categorical variables to Excel}
\usage{
svy_ow(data, var, rounding = 2, hide_prct_char = TRUE, lang = "en")
}
\arguments{
\item{var}{Variable to use}

\item{rounding}{number of digits for rounding (int)}

\item{hide_prct_char}{hides the "\%" characted in the percentages column}

\item{lang}{changes the name of the percentage column (depending on your language). Three values : "en" (default), "fr", or "math"}

\item{Data}{dataframe}
}
\value{
Count table
}
\description{
This functions creates an Excel workbook and exports a oneway table
of multiple variables with counts and percentages
}
