% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_ow_report.r
\name{svy_ow_report}
\alias{svy_ow_report}
\title{Oneway report of multiple categorical variables to Excel, with survey design}
\usage{
svy_ow_report(
  data,
  workbook,
  new_wb = TRUE,
  worksheet,
  vars,
  rounding_n = 0,
  rounding_prct = 2,
  no_na_prct = TRUE,
  cond_prct = FALSE,
  min_cond_prct = 0.05,
  cond_excluded_labels = NULL,
  data_label,
  label_from,
  label_to,
  open_on_finish = TRUE,
  overwrite_file = TRUE,
  lang = "en",
  filename
)
}
\arguments{
\item{workbook}{name of the workbook (string)}

\item{new_wb}{create new workbook (or open filename)}

\item{worksheet}{name of the worksheet (string)}

\item{vars}{Variables to report (string)}

\item{rounding_n}{number of digits for rounding Ns (int)}

\item{rounding_prct}{number of digits for rounding percentages (int)}

\item{no_na_prct}{Removes NAs from \%ages calculations}

\item{cond_prct}{create a new percentage column conditionally, with \% computed conditionally on a threshold (bool)}

\item{min_cond_prct}{Threshold to compute the new \%s (int)}

\item{cond_excluded_labels}{Value labels to be excluded in the calculation of the new \%s, if < min_cond_prct (all if NULL)}

\item{data_label}{Dataframe of value labels from which to retrieve variable labels (optionnal)}

\item{label_from}{Column (in data_label) of variable names}

\item{label_to}{Column (in data_label) of variable labels}

\item{open_on_finish}{open excel file on finish (bool)}

\item{overwrite_file}{Overwrite existing file (bool)}

\item{lang}{Language of header (one of "en", "fr" or "math")}

\item{filename}{Name of excel file to export to (string)}

\item{Data}{dataframe}
}
\value{
Excel file with oneway table
}
\description{
This functions creates an Excel workbook and exports a oneway table
of multiple variables with counts and percentages (with a survey design)
}
