% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_tw.r
\name{svy_tw}
\alias{svy_tw}
\title{Twoway table, with survey design}
\usage{
svy_tw(
  data,
  colvar,
  rowvar,
  type_prct = "col",
  row_total = FALSE,
  rounding_n = 0,
  rounding_prct = 2,
  hide_prct_char = TRUE,
  lang = "en"
)
}
\arguments{
\item{data}{Dataframe to use}

\item{colvar}{Variable to use as the column variable}

\item{rowvar}{Variable to use as the row variable}

\item{type_prct}{Type of percentage shown (one of "col", "row" or "cell")}

\item{row_total}{Add total column for row variable}

\item{rounding_prct}{number of digits for rounding percentages (int)}

\item{hide_prct_char}{hides the "\%" characted in the percentages column}

\item{lang}{changes the name of the percentage column (depending on your language). Three values : "en" (default), "fr", or "math"}
}
\value{
Count table
}
\description{
This functions returns a twoway table of one column variable and one row variable
Must be used on a tbl_svy dataframe
}
